refactor: create reusable ResultList component from address search search result component
