+++ src/lib/focusFirstResult.ts
