+++ src/plugins/addressSearch/components/SearchResults.ce.vue
@@ -1,269 +1,50 @@
 <template>
-	<div
-		v-if="featuresAvailable"
-		id="polar-plugin-address-search-result-wrapper"
-		tabindex="-1"
-	>
-		<template v-for="(result, i) in results" :key="result.categoryId">
-			<span
-				v-if="results.length > 1"
-				:id="`polar-plugin-address-search-category-label-${result.categoryId}`"
-				class="polar-plugin-address-search-category-label"
-			>
-				{{ result.categoryLabel }}
-				{{
-					$t(($) => $.resultCount, {
-						count: searchResults[i].features.features.length,
-						ns: PluginId,
-					})
-				}}
-			</span>
-			<ul
-				:aria-labelledby="`polar-plugin-address-search-category-label-${result.categoryId}`"
-				:class="{
-					'polar-plugin-address-search-list-without-label':
-						results.length === 1,
-				}"
-			>
-				<template
-					v-for="(feature, j) in result.features.features"
-					:key="`result-${i}-${j}`"
-				>
-					<li
-						:id="`polar-plugin-address-search-results-feature-${i}-${j}`"
-						tabindex="-1"
-						@click="addressSearchStore.selectResult(feature, result.categoryId)"
-						@keydown.enter.prevent.stop="
-							addressSearchStore.selectResult(feature, result.categoryId)
-						"
-						@keydown.down.prevent.stop="
-							(event) => focusNextElement(true, event)
-						"
-						@keydown.up.prevent.stop="(event) => focusNextElement(false, event)"
-						@keydown.escape.prevent.stop="escapeResults"
-					>
-						<span class="span-sr-only">{{ feature.title }}</span>
-						<!-- eslint-disable vue/no-v-html -->
-						<span
-							aria-hidden="true"
-							v-html="strongTitleByInput(feature.title, inputValue)"
-						/>
-						<!-- eslint-enable vue/no-v-html -->
-						<component
-							:is="afterResultComponent"
-							v-if="afterResultComponent"
-							:feature="feature"
-						/>
-					</li>
-				</template>
-			</ul>
-			<KernButton
-				v-if="searchResults[i].features.features.length > limitResults"
-				class="kern-btn--tertiary"
-				:icon="
-					areResultsExpanded(result.categoryId)
-						? 'kern-icon--keyboard-arrow-up'
-						: 'kern-icon--keyboard-arrow-down'
-				"
-				@keydown.down.prevent.stop="(event) => focusNextElement(true, event)"
-				@keydown.up.prevent.stop="(event) => focusNextElement(false, event)"
-				@click="toggle(result.categoryId)"
-			>
-				{{
-					$t(
-						($) =>
-							$.resultList[
-								areResultsExpanded(result.categoryId) ? 'reduce' : 'extend'
-							],
-						{ ns: PluginId }
-					)
-				}}
-			</KernButton>
-			<hr
-				v-if="i < results.length - 1"
-				class="kern-divider"
-				aria-hidden="true"
-			/>
-		</template>
-	</div>
+	<PolarResultList
+		:id="PluginId"
+		:features-available="featuresAvailable"
+		:plugin-id="PluginId"
+		:search-results="searchResults"
+		:after-result-component="afterResultComponent ? afterResultComponent : null"
+		:limited-results="limitResults"
+		:input-value="inputValue"
+		:result-count-label="resultCountLabel"
+		:select-result="selectResult"
+		:toggle-label="toggleLabel"
+		:selected-group-id="selectedGroupId"
+		:focus-after-search="addressSearchStore.focusAfterSearch"
+	/>
 </template>
 
 <script setup lang="ts">
-import type { SearchResult } from '../types'
-
+import { t } from 'i18next'
 import { storeToRefs } from 'pinia'
-import { computed, nextTick, ref, toRaw, watch } from 'vue'
 
-import KernButton from '@/components/kern/KernButton.ce.vue'
-import { useCoreStore } from '@/core/stores'
+import PolarResultList from '@/components/PolarResultList.ce.vue'
 
 import { useAddressSearchStore } from '../store'
 import { PluginId } from '../types'
-import { focusFirstResult } from '../utils/focusFirstResult'
-import { strongTitleByInput } from '../utils/strongTitleByInput'
 
-const coreStore = useCoreStore()
 const addressSearchStore = useAddressSearchStore()
+const { selectResult } = addressSearchStore
 const {
 	afterResultComponent,
 	featuresAvailable,
 	inputValue,
 	limitResults,
-	searchResults,
 	selectedGroupId,
+	searchResults,
 } = storeToRefs(addressSearchStore)
 
-// Reset opened categories on group change
-watch(selectedGroupId, () => (openCategories.value = []))
-
-const openCategories = ref<string[]>([])
-
-const results = computed<SearchResult[]>(() =>
-	Array.isArray(addressSearchStore.searchResults)
-		? // If we do not clone, we'd still copy references on the deeper levels
-			structuredClone(toRaw(addressSearchStore.searchResults) as SearchResult[])
-				.filter(
-					(result) => result.groupId === addressSearchStore.selectedGroupId
-				)
-				.reduce<SearchResult[]>((acc, curr) => {
-					const index = acc.findIndex(
-						(val) => val.categoryId === curr.categoryId
-					)
-					if (index === -1) {
-						return [...acc, curr]
-					}
-					;(acc[index] as SearchResult).features.features = [
-						...(acc[index] as SearchResult).features.features,
-						...curr.features.features,
-					]
-
-					return acc
-				}, [])
-				.map((result) => {
-					if (areResultsExpanded(result.categoryId)) {
-						return result
-					}
-
-					result.features.features = result.features.features.slice(
-						0,
-						limitResults.value
-					)
-					return result
-				})
-		: []
-)
-
-watch(results, () => {
-	if (addressSearchStore.focusAfterSearch) {
-		void nextTick(() => {
-			focusFirstResult(results.value.length, coreStore.shadowRoot as ShadowRoot)
-		})
-	}
-})
-
-function areResultsExpanded(category: string) {
-	return openCategories.value.includes(category)
+function resultCountLabel(count: number) {
+	return t(($) => $.resultCount, {
+		count,
+		ns: PluginId,
+	})
 }
 
-function escapeResults() {
-	;(coreStore.shadowRoot as ShadowRoot)
-		.getElementById('polar-plugin-address-search-input')
-		?.focus()
-}
-
-function focusNextElement(down: boolean, event: KeyboardEvent): void {
-	const { target } = event
-
-	if (target === null) {
-		console.warn('Could not focus any element.')
-		return
-	}
-
-	const wrapper = coreStore.shadowRoot?.getElementById(
-		'polar-plugin-address-search-result-wrapper'
-	) as HTMLDivElement
-	const elements = wrapper.querySelectorAll('li, button')
-
-	const index = [...elements].indexOf(target as Element)
-	// Gets the next or previous element in the list of all available results and expansion buttons.
-	const nextElement = elements[(index + (down ? 1 : -1)) % elements.length]
-	if (nextElement) {
-		// @ts-expect-error | we have no non-HTML elements in this DOM part.
-		nextElement.focus()
-		return
-	}
-
-	;(coreStore.shadowRoot as ShadowRoot)
-		.getElementById('polar-plugin-address-search-input')
-		?.focus()
-}
-
-function toggle(category: string) {
-	openCategories.value =
-		openCategories.value.indexOf(category) === -1
-			? [...openCategories.value, category]
-			: openCategories.value.filter((s) => s !== category)
+function toggleLabel(expanded: boolean) {
+	return t(($) => $.resultList[expanded ? 'reduce' : 'extend'], {
+		ns: PluginId,
+	})
 }
 </script>
-
-<style scoped>
-#polar-plugin-address-search-result-wrapper {
-	display: flex;
-	flex-direction: column;
-	gap: var(--kern-metric-space-2x-small);
-	width: 100%;
-	padding-bottom: 0.625rem;
-	overflow-y: auto;
-
-	.polar-plugin-address-search-category-label {
-		display: flex;
-		align-items: center;
-		min-height: var(--kern-metric-dimension-large);
-		padding: 0 var(--kern-metric-space-small);
-		margin: 0;
-		font-size: calc(var(--kern-typography-font-size-small-static) * 0.875);
-		font-weight: normal;
-		color: var(--kern-color-layout-text-muted);
-	}
-
-	.polar-plugin-address-search-list-without-label {
-		margin-top: var(--kern-metric-space-x-small);
-	}
-
-	ul {
-		margin: 0;
-		padding: 0;
-
-		li {
-			display: flex;
-			align-items: center;
-			min-height: var(--kern-metric-dimension-x-large);
-			padding: var(--kern-metric-space-none) var(--kern-metric-space-small);
-			margin: var(--kern-metric-space-none) var(--kern-metric-space-small);
-			border-radius: var(--kern-metric-border-radius-default);
-			color: var(--kern-color-layout-text-default);
-			transition: 0.3s cubic-bezier(0.25, 0.8, 0.5, 1);
-
-			&:hover,
-			&:focus {
-				background-color: var(--kern-color-layout-background-hued);
-				cursor: pointer;
-			}
-		}
-	}
-
-	button {
-		margin: var(--kern-metric-space-none) var(--kern-metric-space-small);
-	}
-}
-/* Copy of kern-sr-only with a normal height so screen reader focus is correct */
-.span-sr-only {
-	width: 1px;
-	padding: 0;
-	margin: -1px;
-	overflow: hidden;
-	clip-path: circle(0);
-	white-space: nowrap;
-	border: 0;
-}
-</style>
